backend/fastapi-api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.35
pydantic>=2.10.0
python-dotenv>=1.0.1
